feat : 인식 결과 문자열 DB에 저장
